+++ bp-better-messages/addons/peepso-groups.php
@@ -423,6 +423,10 @@ if ( !class_exists( 'Better_Messages_Peepso_Groups' ) ) {
 
             if( count($recipients) > 0 ) {
                 foreach ($recipients as $user_id => $recipient) {
+                    if( $user_id < 0 ){
+                        continue;
+                    }
+
                     global $wpdb;
 
                     $wpdb->delete( bm_get_table('recipients'), [
